requirements.txt
guidance==0.1.10
llama-cpp-python==0.2.36